Bug fixes and improvements in code

+++ main.py
@@ -2,35 +2,6 @@ import pandas as pd
 from datetime import date
 
 
-def addNewBook():
-    # Get the details of the new book from user input
-    bookid = int(input("Enter a book id: "))
-    title = input("Enter book title: ")
-    author = input("Enter author of the book: ")
-    publisher = input("Enter book publisher: ")
-    edition = input("Enter edition of book: ")
-    cost = float(input("Enter cost of the book: "))
-    category = input("Enter category of book: ")
-
-    # Load the current books data from the CSV file
-    bdf = pd.read_csv("books.csv")
-
-    # Create a new DataFrame for the new book
-    new_book = pd.DataFrame([[bookid, title, author, publisher, edition, cost, category]],
-                            columns=["bookid", "title", "author", "publisher", "edition", "cost", "category"])
-
-    # Concatenate the new book to the existing DataFrame
-    bdf = pd.concat([bdf, new_book], ignore_index=True)
-
-    # Save the updated DataFrame back to the CSV file
-    bdf.to_csv("books.csv", index=False)
-
-    print("Book added successfully")
-    print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")
-    prompt()
-
-
-
 def login():
     uname = input("Enter Username : ")
     pwd = input("Enter Password : ")
@@ -48,32 +19,72 @@ def login():
             print("Username and Password matched successfully")
             return True
 
+
+def addNewBook():
+    # Get the details of the new book from user input
+    bookid = int(input("Enter a book id: "))
+    title = input("Enter book title: ")
+    author = input("Enter author of the book: ")
+    publisher = input("Enter book publisher: ")
+    edition = input("Enter edition of book: ")
+    cost = float(input("Enter cost of the book: "))
+    category = input("Enter category of book: ")
+
+    # Load the current books data from the CSV file
+    bdf = pd.read_csv("books.csv")
+
+    # Create a new DataFrame for the new book
+    new_book = pd.DataFrame([[bookid, title.lower(), author, publisher, edition, cost, category]],
+                            columns=["bookid", "title", "author", "publisher", "edition", "cost", "category"])
+
+    # Concatenate the new book to the existing DataFrame
+    bdf = pd.concat([bdf, new_book], ignore_index=True)
+
+    # Save the updated DataFrame back to the CSV file
+    bdf.to_csv("books.csv", index=False)
+
+    print("Book added successfully\n")
+    print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")
+    prompt()
+
+
+
+
 def searchBook():
     name = input("Enter book title to be searched : ")
     bdf = pd.read_csv("books.csv")
-    df = bdf.loc[bdf["title"] == name]
+    df = bdf.loc[bdf["title"] == name.lower()]
     if df.empty:
-        print("No book found with given title")
+        print("No book found with given title\n")
         print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")
         prompt()
     else:
-        print("Book details are ")
+        print("Book details are : ")
         print(df)
+        print("\n")
         print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")
         prompt()
 
 def deleteBook():
     name = input("Enter book title to be deleted : ")
     bdf = pd.read_csv("books.csv")
-    bdf = bdf.drop(bdf[bdf["title"] == name].index)
-    bdf.to_csv("books.csv",index = False)
-    print("Book Deleted Successfully")
-    print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")
-    prompt()
+    tempdf = bdf.loc[bdf["title"] == name]
+
+    if tempdf.empty:
+        print("There is no book of such name in the library. Please try again\n")
+        print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")
+        prompt()
+    else:
+        bdf = bdf.drop(bdf[bdf["title"] == name.lower()].index)
+        bdf.to_csv("books.csv",index = False)
+        print("Book DELETED Successfully\n")
+        print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")
+        prompt()
 
 def showBooks():
     bdf = pd.read_csv("books.csv")
     print(bdf)
+    print("\n")
     print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")
     prompt()
 
@@ -81,7 +92,7 @@ def showBooks():
 def addNewMember():
     # Get the details of the new member from user input
     mid = int(input("Enter Member id: "))
-    name = input("Enter name of the member: ")
+    name = input("Enter name of the member: ").lower()
     phone = input("Enter phone number: ")
     email = input("Enter email id: ")
     address = input("Enter address: ")
@@ -98,62 +109,71 @@ def addNewMember():
     # Save the updated DataFrame back to the CSV file
     mdf.to_csv("members.csv", index=False)
 
-    print("Member added successfully")
+    print("Member added successfully\n")
     print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")
     prompt()
 
 
 
 def searchMember():
-    name = input("Enter member name to be searched : ")
+    name = input("Enter member name to be searched : ").lower()
     mdf = pd.read_csv("members.csv")
     df = mdf.loc[mdf["name"] == name]
     if df.empty:
-        print("No member found with given name")
+        print("No member found with given name\n")
         print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")
         prompt()
     else:
-        print("Member details are ")
+        print("Member details are: ")
         print(df)
+        print("\n")
         print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")
         prompt()
 
 
 def deleteMember():
-    name = input("Enter member name to be deleted : ")
+    name = input("Enter member name to be deleted : ").lower()
     mdf = pd.read_csv("members.csv")
-    mdf = mdf.drop(mdf[mdf["name"] == name].index)
-    mdf.to_csv("members.csv", index=False)
-    print("Member Deleted Successfully")
-    print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")
-    prompt()
+    tempdf = mdf.loc[mdf["name"] == name]
+
+    if tempdf.empty:
+        print("There is no member of such name registered. Please try again\n")
+        print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")
+        prompt()
+    else:
+        mdf = mdf.drop(mdf[mdf["name"] == name].index)
+        mdf.to_csv("members.csv", index=False)
+        print("Member DELETED Successfully\n")
+        print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")
+        prompt()
 
 def showMembers():
     mdf = pd.read_csv("members.csv")
     print(mdf)
+    print("\n")
     print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")
     prompt()
 
 
 def issueBook():
-    bname = input("Enter Book name to be searched : ")
+    bname = input("Enter Book name to be searched : ").lower()
     df = pd.read_csv("books.csv")
     df = df.loc[df["title"] == bname]
     if df.empty:
-        print("No Book Found in the Library")
+        print("No Book Found in the Library\n")
         print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")   
         prompt()
 
-    mname = input("Enter member name to be searched : ")
+    mname = input("Enter member name to be searched : ").lower()
     df = pd.read_csv("members.csv")
     df = df.loc[df["name"] == mname]  # Changed 'name' to 'Name' to match the column name
     if df.empty:
-        print("No such Member Found")
+        print("No such Member Found\n")
         print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")
         prompt()
 
     idf = pd.read_csv("issuedbooks.csv")
-    book_issue = [bname, mname, date.today(), ""]
+    book_issue = [bname, mname, date.today()]
 
     # Use len(idf) to get the index for the next available row
     n = len(idf)  
@@ -163,7 +183,7 @@ def issueBook():
 
     # Save the updated DataFrame back to the issuedbooks.csv file
     idf.to_csv("issuedbooks.csv", index=False)
-    print("Book Issued Successfully")
+    print("Book Issued Successfully\n")
     print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")
     prompt()
 
@@ -171,37 +191,38 @@ def issueBook():
 def showIssuedBooks():
     idf = pd.read_csv("issuedbooks.csv")
     print(idf)
+    print("\n")
     print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")
     prompt()
 
 
 def returnBook():
-    bname = input("Enter Book to be returned : ")
-    mname = input("Enter Member who has the book : ")
+    bname = input("Enter Book to be returned : ").lower()
+    mname = input("Enter Member who has the book : ").lower()
     idf = pd.read_csv("issuedbooks.csv")
     idf = idf.loc[idf["book_name"] == bname]
     if idf.empty:
-        print("The book is not issued in records")
+        print("The book is not issued in records\n")
         print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")
         prompt()
     else:
         idf = idf.loc[idf["member_name"] == mname]
         if idf.empty:
-            print("The book is not issued to the member")
+            print("The book is not issued to the member\n")
             print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")    
             prompt()
         else:
-            print("Book can be returned")
+            print("Book can be returned\n")
             ans = input("Are you sure you want to return the book : ")
             if ans.lower() == "yes":
                 idf = pd.read_csv("issuedbooks.csv")
                 idf = idf.drop(idf[idf["book_name"] == bname].index)
                 idf.to_csv("issuedbooks.csv", index=False)
-                print("Book Returned Successfully")
+                print("Book Returned Successfully\n")
                 print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")
                 prompt()
             else:
-                print("Return operation cancelled")
+                print("Return operation cancelled.\n")
                 print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")              
                 prompt()
 
@@ -209,18 +230,18 @@ def showMenu():
     print("-----------------------------")
     print("         MPS LIBRARY         ")
     print("-----------------------------")
-    print("Press 1 - Add a New Book")
+    print("Press 1 - Add New Book")
     print("Press 2 - Search for a Book")
-    print("Press 3 - Delete a Book")
+    print("Press 3 - Delete Book")
     print("Press 4 - Show All Books")
-    print("Press 5 - Add a New Member")
+    print("Press 5 - Add New Member")
     print("Press 6 - Search for a Member")
-    print("Press 7 - Delete a Member")
+    print("Press 7 - Delete Member")
     print("Press 8 - Show All Members")
     print("Press 9 - Issue a Book")
     print("Press 10 - Return a Book")
-    print("Press 11 - Show All Issued Books")
-    print("Press 12 - To Exit")
+    print("Press 11 - Show Issuing Records")
+    print("Press 12 - To Quit")
     choice = input("Enter your choice : ")
     return choice
 
@@ -230,9 +251,9 @@ def showMenu():
 
 def prompt():
     if login():
-
-        ans = input("Do you wish to perform any other tasks?")
-        if ans == "yes":
+        print('\n')
+        ans = input("Do you wish to perform any other tasks?: ").lower()
+        if ans == "yes" or ans == "y":
             ch = showMenu()
             if ch == '1':
                 addNewBook()
@@ -259,13 +280,13 @@ def prompt():
             elif ch == '12':
                 return 
             else:
-               print("Invalid Option Selected")
+               print("Invalid Option Selected\n")
                print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")              
                prompt()
-        elif ans == "no":
+        elif ans == "no" or ans == "n":
             return
         else:
-            print("invalid input recieved. Try running the program if you want to..")
+            print("invalid input recieved. Try running the program if you want to..\n")
             print("ENTER YOUR LOGIN DETAILS AGAIN TO ACCESS THE PROGRAM AGAIN........")              
             prompt()
